monitoring-backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv
